# docs/swagger.yaml
basePath: /
definitions:
  controllers.ErrorResponse:
    properties:
      error:
        description: Описание возникшей ошибки...
        type: string
    type: object
  controllers.accessTokenResponse:
    properties:
      access_token:
        description: JWT токен для аутентификации пользователя...
        type: string
    type: object
...
  models.SignInInput:
    properties:
      password:
        type: string
      username:
        type: string
    type: object
...
host: localhost:8585
info:
  contact: {}
  description: API Server for eShop Application
  title: eShop API
  version: "1.02"
paths:
  /auth/sign-in:
    post:
      consumes:
      - application/json
      description: User authentication (returns JWT token)
      operationId: sign-in
      parameters:
      - description: Login data
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/models.SignInInput'
      produces:
      - application/json
      responses:
        "200":
          description: access_token
          schema:
            $ref: '#/definitions/controllers.accessTokenResponse'
        "400":
          description: Invalid input
          schema:
            $ref: '#/definitions/controllers.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/controllers.ErrorResponse'
        "500":
          description: Server error
          schema:
            $ref: '#/definitions/controllers.ErrorResponse'
      summary: Log in user
      tags:
      - auth
...
securityDefinitions:
  ApiKeyAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"